Create Password generator.py
Python Code to create a Random Password Generator. In this Random Password Generator, the user will be required to enter How many letters, symbols and numbers user want in his password after which he'll be provided with a random password. Here, we'll be using the 'random' module in Python to create our project